Move the public path for `channel_state` types to their new path
